requirements.txt
django>1.6,<1.7
pillow==2.3.0
djangorestframework>2.3.12,<2.4
Markdown==2.4
